consolidate files in core/data
