--- nimble/core/data/dataframeFeatures.py
@@ -1,77 +1,0 @@
-"""
-Method implementations and helpers acting specifically on features in a
-DataFrame object.
-"""
-
-import numpy
-
-from nimble._utility import pd
-from .axis_view import AxisView
-from .dataframeAxis import DataFrameAxis
-from .features import Features
-from .features_view import FeaturesView
-
-class DataFrameFeatures(DataFrameAxis, Features):
-    """
-    DataFrame method implementations performed on the feature axis.
-
-    Parameters
-    ----------
-    base : DataFrame
-        The DataFrame instance that will be queried and modified.
-    """
-
-    ##############################
-    # Structural implementations #
-    ##############################
-
-    def _insert_implementation(self, insertBefore, toInsert):
-        """
-        Insert the features from the toInsert object to the right of the
-        provided index in this object, the remaining points from this
-        object will continue to the right of the inserted points.
-        """
-        startData = self._base.data.iloc[:, :insertBefore]
-        endData = self._base.data.iloc[:, insertBefore:]
-        self._base.data = pd.concat((startData, toInsert.data, endData),
-                                    axis=1, ignore_index=True)
-
-    def _transform_implementation(self, function, limitTo):
-        for j, f in enumerate(self):
-            if limitTo is not None and j not in limitTo:
-                continue
-            currRet = function(f)
-
-            self._base.data.iloc[:, j] = currRet
-
-
-    ################################
-    # Higher Order implementations #
-    ################################
-
-    def _splitByParsing_implementation(self, featureIndex, splitList,
-                                       numRetFeatures, numResultingFts):
-        tmpData = numpy.empty(shape=(len(self._base.points), numRetFeatures),
-                              dtype=numpy.object_)
-
-        tmpData[:, :featureIndex] = self._base.data.values[:, :featureIndex]
-        for i in range(numResultingFts):
-            newFeat = []
-            for lst in splitList:
-                newFeat.append(lst[i])
-            tmpData[:, featureIndex + i] = newFeat
-        existingData = self._base.data.values[:, featureIndex + 1:]
-        tmpData[:, featureIndex + numResultingFts:] = existingData
-
-        self._base.data = pd.DataFrame(tmpData)
-
-class DataFrameFeaturesView(FeaturesView, AxisView, DataFrameFeatures):
-    """
-    Limit functionality of DataFrameFeatures to read-only.
-
-    Parameters
-    ----------
-    base : DataFrameView
-        The DataFrameView instance that will be queried.
-    """
-    pass
