+++ nimble/core/data/listAxis.py
@@ -10,7 +10,14 @@ import numpy
 
 import nimble
 from .axis import Axis
+from .views import AxisView
+from .points import Points
+from .views import PointsView
+from .features import Features
+from .views import FeaturesView
 from ._dataHelpers import denseAxisUniqueArray, uniqueNameGetter
+from ._dataHelpers import fillArrayWithCollapsedFeatures
+from ._dataHelpers import fillArrayWithExpandedFeatures
 
 class ListAxis(Axis):
     """
@@ -136,3 +143,174 @@ class ListAxis(Axis):
     @abstractmethod
     def _transform_implementation(self, function, limitTo):
         pass
+
+
+class ListPoints(ListAxis, Points):
+    """
+    List method implementations performed on the points axis.
+
+    Parameters
+    ----------
+    base : List
+        The List instance that will be queried and modified.
+    """
+
+    ##############################
+    # Structural implementations #
+    ##############################
+
+    def _insert_implementation(self, insertBefore, toInsert):
+        """
+        Insert the points from the toInsert object below the provided
+        index in this object, the remaining points from this object will
+        continue below the inserted points.
+        """
+        insert = toInsert.copy('List').data
+        if insertBefore != 0 and insertBefore != len(self):
+            start = self._base.data[:insertBefore]
+            end = self._base.data[insertBefore:]
+            allData = start + insert + end
+        elif insertBefore == 0:
+            allData = insert + self._base.data
+        else:
+            allData = self._base.data + insert
+
+        self._base.data = allData
+
+    def _transform_implementation(self, function, limitTo):
+        for i, p in enumerate(self):
+            if limitTo is not None and i not in limitTo:
+                continue
+            currRet = function(p)
+
+            self._base.data[i] = list(currRet)
+
+    ################################
+    # Higher Order implementations #
+    ################################
+
+    def _splitByCollapsingFeatures_implementation(
+            self, featuresToCollapse, collapseIndices, retainIndices,
+            currNumPoints, currFtNames, numRetPoints, numRetFeatures):
+        collapseData = []
+        retainData = []
+        for pt in self._base.data:
+            collapseFeatures = []
+            retainFeatures = []
+            for idx in collapseIndices:
+                collapseFeatures.append(pt[idx])
+            for idx in retainIndices:
+                retainFeatures.append(pt[idx])
+            collapseData.append(collapseFeatures)
+            retainData.append(retainFeatures)
+
+        tmpData = fillArrayWithCollapsedFeatures(
+            featuresToCollapse, retainData, numpy.array(collapseData),
+            currNumPoints, currFtNames, numRetPoints, numRetFeatures)
+
+        self._base.data = tmpData.tolist()
+        self._base._numFeatures = numRetFeatures
+
+    def _combineByExpandingFeatures_implementation(self, uniqueDict, namesIdx,
+                                                   uniqueNames, numRetFeatures,
+                                                   numExpanded):
+        tmpData = fillArrayWithExpandedFeatures(uniqueDict, namesIdx,
+                                                uniqueNames, numRetFeatures,
+                                                numExpanded)
+
+        self._base.data = tmpData.tolist()
+        self._base._numFeatures = numRetFeatures
+
+
+class ListPointsView(PointsView, AxisView, ListPoints):
+    """
+    Limit functionality of ListPoints to read-only.
+
+    Parameters
+    ----------
+    base : ListView
+        The ListView instance that will be queried.
+    """
+    pass
+
+
+class ListFeatures(ListAxis, Features):
+    """
+    List method implementations performed on the feature axis.
+
+    Parameters
+    ----------
+    base : List
+        The List instance that will be queried and modified.
+    """
+
+    ##############################
+    # Structural implementations #
+    ##############################
+
+    def _insert_implementation(self, insertBefore, toInsert):
+        """
+        Insert the features from the toInsert object to the right of the
+        provided index in this object, the remaining points from this
+        object will continue to the right of the inserted points.
+        """
+        insert = toInsert.copy('pythonlist')
+        if insertBefore != 0 and insertBefore != len(self):
+            breakIdx = insertBefore - 1
+            restartIdx = insertBefore
+            start = self._base.view(featureEnd=breakIdx).copy('pythonlist')
+            end = self._base.view(featureStart=restartIdx).copy('pythonlist')
+            zipChunks = zip(start, insert, end)
+            allData = list(map(lambda pt: pt[0] + pt[1] + pt[2], zipChunks))
+        elif insertBefore == 0:
+            end = self._base.copy('pythonlist')
+            allData = list(map(lambda pt: pt[0] + pt[1], zip(insert, end)))
+        else:
+            start = self._base.copy('pythonlist')
+            allData = list(map(lambda pt: pt[0] + pt[1], zip(start, insert)))
+
+        self._base.data = allData
+        self._base._numFeatures += len(toInsert.features)
+
+    def _transform_implementation(self, function, limitTo):
+        for j, f in enumerate(self):
+            if limitTo is not None and j not in limitTo:
+                continue
+            currRet = function(f)
+
+            for i in range(len(self._base.points)):
+                self._base.data[i][j] = currRet[i]
+
+    ################################
+    # Higher Order implementations #
+    ################################
+
+    def _splitByParsing_implementation(self, featureIndex, splitList,
+                                       numRetFeatures, numResultingFts):
+        tmpData = numpy.empty(shape=(len(self._base.points), numRetFeatures),
+                              dtype=numpy.object_)
+
+        tmpData[:, :featureIndex] = [ft[:featureIndex] for ft
+                                     in self._base.data]
+        for i in range(numResultingFts):
+            newFeat = []
+            for lst in splitList:
+                newFeat.append(lst[i])
+            tmpData[:, featureIndex + i] = newFeat
+        existingData = [ft[featureIndex + 1:] for ft in self._base.data]
+        tmpData[:, featureIndex + numResultingFts:] = existingData
+
+        self._base.data = tmpData.tolist()
+        self._base._numFeatures = numRetFeatures
+
+
+class ListFeaturesView(FeaturesView, AxisView, ListFeatures):
+    """
+    Limit functionality of ListFeatures to read-only.
+
+    Parameters
+    ----------
+    base : ListView
+        The ListView instance that will be queried.
+    """
+    pass
